Move LintModuleOutputUpdate to pylint.testutil.functional
