+++ pylint/testutils/functional/__init__.py
@@ -7,12 +7,14 @@ __all__ = [
     "get_functional_test_files_from_directory",
     "NoFileError",
     "parse_python_version",
+    "LintModuleOutputUpdate",
 ]
 
 from pylint.testutils.functional.find_functional_tests import (
     REASONABLY_DISPLAYABLE_VERTICALLY,
     get_functional_test_files_from_directory,
 )
+from pylint.testutils.functional.lint_module_output_update import LintModuleOutputUpdate
 from pylint.testutils.functional.test_file import (
     FunctionalTestFile,
     NoFileError,
