+++ tests/test_functional.py
@@ -22,8 +22,6 @@
 # For details: https://github.com/PyCQA/pylint/blob/main/LICENSE
 
 """Functional full-module tests for PyLint."""
-import csv
-import os
 import sys
 from pathlib import Path
 from typing import Union
@@ -38,6 +36,7 @@ from pylint.testutils.functional import (
     FunctionalTestFile,
     get_functional_test_files_from_directory,
 )
+from pylint.testutils.functional.lint_module_output_update import LintModuleOutputUpdate
 from pylint.utils import HAS_ISORT_5
 
 # TODOs
@@ -45,26 +44,6 @@ from pylint.utils import HAS_ISORT_5
 
 
 FUNCTIONAL_DIR = Path(__file__).parent.resolve() / "functional"
-
-
-class LintModuleOutputUpdate(testutils.LintModuleTest):
-    """If message files should be updated instead of checked."""
-
-    class TestDialect(csv.excel):
-        delimiter = ":"
-        lineterminator = "\n"
-
-    csv.register_dialect("test", TestDialect)
-
-    def _check_output_text(self, _, expected_output, actual_output):
-        if not expected_output and not actual_output:
-            if os.path.exists(self._test_file.expected_output):
-                os.remove(self._test_file.expected_output)
-            return
-        with open(self._test_file.expected_output, "w", encoding="utf-8") as f:
-            writer = csv.writer(f, dialect="test")
-            for line in actual_output:
-                writer.writerow(line.to_csv())
 
 
 # isort 5 has slightly different rules as isort 4. Testing both would be hard: test with isort 5 only.
